Move TreeChangeNotification to core
